Add tmux and fix startify.
Can't believe I was had my startify setup so badly. Now it no longer
changes to CVS root, and it doesn't change to the directory of the file
either.

Also add a tmux mapping (thanks to @minimul ->
http://minimul.com/increased-developer-productivity-with-tmux-part-5.html)
to simply send the up arrow to the right hand pane. Easy way to re-run
a command.