Allow a default for gmr
`gmr` shows the most recent branches used in a repo. I seem to have
quite a few branches in some repositories, and it can flow off the page.

This change defaults to only showing the most recent 20 branches, but,
you can instead call it like `gmr 30` to show the 30 most recent branches.