Add timestamp to history files.
Seems like a useful thing to have. I was ending up with weird
duplication (which I imagine has been there for a long time). Removing
the PROMPT_COMMAND seems to fix this. Probably some hack I copied from
somewhere without really understanding it.

which is not to mean I understand it now of course ;) only that removing
it seems to fix the duplicated history issue. So, empirically it's
better anyway ;)